% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aSPUscore.R
\name{aSPU}
\alias{aSPU}
\title{Sum of Powered Score (SPU) tests and adaptive SPU (aSPU) test for single trait - SNP set association.}
\usage{
aSPU(U, V, pow = c(1:8, Inf), n.perm = 1000)
}
\arguments{
\item{U}{Score vector for the marker set we are interested in.}

\item{V}{Corresponding covariance matrix for the score vector.}

\item{pow}{power used in SPU test. A vector of the powers.}

\item{n.perm}{number of permutations or bootstraps.}
}
\value{
A list object, Ts : test statistics for the SPU tests (in the order of the specified pow) and finally for the aSPU test.
        pvs : p-values for the SPU and aSPU tests.
}
\description{
It gives p-values of the SPU tests and aSPU test.
}
\author{
Chong Wu and Wei Pan
}
\references{
Wei Pan, Junghi Kim, Yiwei Zhang, Xiaotong Shen and Peng Wei (2014)
A powerful and adaptive association test for rare variants,
Genetics, 197(4), 1081-95
}

